--- calculator-data/src/main/java/camp/nextstep/edu/calculator/ExpressionListTypeConverter.kt
@@ -1,13 +1,0 @@
-package camp.nextstep.edu.calculator
-
-import androidx.room.TypeConverter
-import com.google.gson.Gson
-
-class ExpressionListTypeConverter {
-    @TypeConverter
-    fun listToJson(value: List<CalculatorResultData>?) = Gson().toJson(value)
-
-    @TypeConverter
-    fun jsonToList(value: String) = Gson().fromJson(value, Array<CalculatorResultData>::class.java).toList()
-
-}
